@@ -4,6 +4,7 @@
   "description": "Using Javascript to parse Wiktionary Files -- A work in progress",
   "main": "main.js",
   "scripts": {
+    "tslint": "find . \\( -path './visapp*.ts*' -or -path './test*.ts*' \\) -exec ./node_modules/.bin/tslint {} \\;",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -20,36 +21,46 @@
     "async": "^1.5.0",
     "bluebird": "^3.1.1",
     "classnames": "^2.2.3",
+    "detect-node": "^2.0.3",
     "es6-promise": "^3.0.2",
     "express": "^4.13.3",
     "history": "^1.17.0",
     "html-entities": "^1.2.0",
     "immutable": "^3.7.6",
     "install": "^0.4.1",
+    "isomorphic-fetch": "^2.2.1",
+    "jQuery": "^1.7.4",
     "jison": "^0.4.16",
-    "jquery": "^2.2.0",
+    "jquery": "^2.2.1",
+    "jsdom": "^8.1.0",
     "jsonpath": "^0.2.2",
     "lodash": "^3.10.1",
     "npm": "^3.5.2",
     "prettyjson": "^1.1.3",
-    "react": "^0.14.6",
+    "react": "^0.14.7",
     "react-dom": "^0.14.6",
     "react-redux": "^4.0.6",
     "react-router": "^1.0.3",
-    "redux": "^3.0.5",
+    "react-router-redux": "^2.1.0",
+    "redux": "^3.3.1",
+    "redux-devtools": "^3.0.2",
     "redux-simple-router": "^2.0.4",
     "request": "^2.67.0",
     "request-promise": "^1.0.2",
-    "rx": "^4.0.7",
+    "rx": "^4.1.0",
+    "rx-dom": "^7.0.3",
     "rxjs": "^5.0.0-beta.1",
     "systemjs": "^0.19.9",
-    "typescript": "^1.7.5"
+    "typescript": "^1.7.5",
+    "whatwg-fetch": "^0.11.0"
   },
   "devDependencies": {
     "JSONSelect": "^0.4.0",
     "chai": "^3.4.1",
     "html-webpack-plugin": "^1.7.0",
     "mocha": "^2.3.4",
+    "tslint": "^3.6.0",
+    "typescript": "^1.8.7",
     "webpack": "^1.12.11"
   }
 }
