@@ -2,16 +2,15 @@
  * Created by jasondent on 23/01/2016.
  */
 
-/// <reference path="../../typings/tsd.d.ts" />
-/// <reference path="./../interfaces.d.ts"/>
-
 // Actions needed to be added in two places:
 // 1 as a constant and 2 in the Actions.
 const REQUEST_PAGE = 'REQUEST_PAGE';
 const APPLY_AST = 'APPLY_AST';
+const REQUEST_PAGE_READY = 'REQUEST_PAGE_READY';
 
 export const Actions = {
     REQUEST_PAGE,
+    REQUEST_PAGE_READY,
     APPLY_AST
 };
 
@@ -25,12 +24,17 @@ export interface Action {
 }
 
 export interface PageRequest extends PayLoad { lang: string, page: string, site: string }
-export function requestPage(request: PageRequest) : Action {
-    return { type: Actions.REQUEST_PAGE, payload: request };
+export function requestPage(payload: PageRequest) : Action {
+    return { type: Actions.REQUEST_PAGE, payload };
 }
 
 
-export function applyAst(ast: IAstModel) {
-    return { type: Actions.APPLY_AST, payload: ast };
+export interface  RequestPageReady extends PayLoad { lang: string, page: string, site: string, ast: IAstModel }
+export function requestPageReady(payload: RequestPageReady) : Action {
+    return { type: Actions.REQUEST_PAGE_READY, payload };
+}
+
+export function applyAst() {
+    return { type: Actions.APPLY_AST, payload:{} };
 }
 
